@@ -2,7 +2,7 @@ name: pill_pusher
 description: a reminder app
 
 # The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
+# pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
@@ -18,35 +18,36 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.18.2 <3.0.0"
 
+# Dependencies specify other packages that your package needs in order to work.
+# To automatically upgrade your package dependencies to the latest versions
+# consider running `flutter pub upgrade --major-versions`. Alternatively,
+# dependencies can be manually updated by changing the version numbers below to
+# the latest version available on pub.dev. To see which dependencies have newer
+# versions available, run `flutter pub outdated`.
 dependencies:
-  flutter:
-    sdk: flutter
-  # Service locator
-  get_it: ^4.0.2
-  # Bloc for state management
-#  flutter_bloc: ^5.0.1
-  # Value equality
-  equatable: ^1.2.2
-  # Functional programming thingies
-  dartz: ^0.9.1
-  # Remote API
-  connectivity: ^0.4.9
-  http: ^0.12.2
-  # Local cache
-  shared_preferences: ^0.5.8
-
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.5
+  connectivity: ^3.0.6
+  dartz: ^0.10.1
+  equatable: ^2.0.5
+  flutter:
+    sdk: flutter
+  get_it: ^7.2.0
+  #  flutter_bloc: ^5.0.1
+  http: ^0.13.5
+  shared_preferences: ^2.0.15
+  very_good_analysis: ^3.1.0
+
 
 dev_dependencies:
-  test: ^1.14.3
+  dart_code_metrics: ^5.0.1
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.1
+  mockito: ^5.3.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
