@@ -1,4 +1,3 @@
-import 'package:meta/meta.dart';
 import 'package:equatable/equatable.dart';
 import 'package:pill_pusher/domain/entities/pill.dart';
 
@@ -9,11 +8,10 @@ class PillSet extends Equatable {
 
   //TODO this probably should be a named constructor to enable making fields private
   PillSet({
-    @required this.name,
-    this.frequency : '',
-    this.pills : const [],
+    required this.name,
+    this.frequency: '',
+    this.pills: const [],
   });
-
 
   @override
   List<Object> get props => [name, frequency, pills];
@@ -22,9 +20,7 @@ class PillSet extends Equatable {
   bool get stringify => true;
 
   factory PillSet.fromJson(Map<String, dynamic> json) {
-    var somePills = json['pills'].map((pill) => Pill.fromJson(pill))
-        .toList()
-        .cast<Pill>();
+    var somePills = json['pills'].map((pill) => Pill.fromJson(pill)).toList().cast<Pill>();
 
     var pillSet = PillSet(
       name: json['name'],
@@ -33,7 +29,6 @@ class PillSet extends Equatable {
     );
 
     return pillSet;
-
   }
 
   // TODO add tests for this!
