@@ -1,21 +1,11 @@
-import 'package:meta/meta.dart';
-import 'pill_box_model.dart';
-
-import 'pill_model.dart';
-import '../../domain/entities/pill_box_set.dart';
+import 'package:pill_pusher/features/schedule/data/models/pill_box_model.dart';
+import 'package:pill_pusher/features/schedule/domain/entities/pill_box_set.dart';
 
 class PillBoxSetModel extends PillBoxSet {
   final List<PillBoxModel> pillBoxes;
 
-  PillBoxSetModel({
-    @required String caretaker,
-    @required String dependent,
-    @required this.pillBoxes
-  }) : super(
-      caretaker: caretaker,
-      dependent: dependent,
-    pillBoxes: pillBoxes
-  );
+  PillBoxSetModel({required String caretaker, required String dependent, required this.pillBoxes})
+      : super(caretaker: caretaker, dependent: dependent, pillBoxes: pillBoxes);
 
   factory PillBoxSetModel.fromJson(Map<String, dynamic> json) {
     return PillBoxSetModel(
@@ -23,8 +13,8 @@ class PillBoxSetModel extends PillBoxSet {
       dependent: json['dependent'],
       pillBoxes: json['pillBoxes']
           .map((pillBox) => PillBoxModel.fromJson(pillBox))
-            .toList()
-            .cast<PillBoxModel>(),
+          .toList()
+          .cast<PillBoxModel>(),
     );
   }
 
