@@ -1,10 +1,8 @@
-import 'package:meta/meta.dart';
-
-import '../../domain/entities/pill.dart';
+import 'package:pill_pusher/features/schedule/domain/entities/pill.dart';
 
 class PillModel extends Pill {
   PillModel({
-    @required String name,
+    required String name,
   }) : super(name: name);
 
   factory PillModel.fromJson(Map<String, dynamic> json) {
