@@ -1,11 +1,10 @@
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 
 class Pill extends Equatable {
   final String name;
 
   Pill({
-    @required this.name,
+    required this.name,
   });
 
   @override
