@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
+
 import 'pill_box.dart';
 
 class PillBoxSet extends Equatable {
@@ -7,11 +7,7 @@ class PillBoxSet extends Equatable {
   final String dependent;
   final List<PillBox> pillBoxes;
 
-  PillBoxSet({
-    @required this.caretaker,
-    @required this.dependent,
-    @required this.pillBoxes
-  });
+  PillBoxSet({required this.caretaker, required this.dependent, required this.pillBoxes});
 
   @override
   List<Object> get props => [caretaker, dependent, pillBoxes];
