@@ -1,12 +1,10 @@
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 
+import '../../../../core/error/failure.dart';
+import '../../../../core/usecases/usecase.dart';
 import '../entities/pill_box_set.dart';
 import '../repositories/pill_box_set_repository.dart';
-import '../../../../core/usecases/usecase.dart';
-import '../../../../core/error/failure.dart';
-
 
 class GetPillBoxSet implements UseCase<PillBoxSet, Params> {
   final PillBoxSetRepository repository;
@@ -22,7 +20,7 @@ class GetPillBoxSet implements UseCase<PillBoxSet, Params> {
 class Params extends Equatable {
   final String dependent;
 
-  Params({@required this.dependent});
+  Params({required this.dependent});
 
   @override
   List<Object> get props => [dependent];
