@@ -5,6 +5,8 @@ abstract class Failure extends Equatable {
   List<Object> get props => [];
 }
 
-class ServerFailure extends Failure {}
-
 class CacheFailure extends Failure {}
+
+class NotFoundFailure extends Failure {}
+
+class ServerFailure extends Failure {}
