@@ -1,23 +1,23 @@
-import 'dart:convert';
-
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/mockito.dart';
 import 'package:http/http.dart' as http;
+import 'package:mockito/mockito.dart';
 import 'package:pill_pusher/features/schedule/data/datasources/pill_box_set_remote_data_source.dart';
+
 import '../../../../fixtures/fixture_reader.dart';
 
 class MockHttpClient extends Mock implements http.Client {}
 
 void main() {
-  PillBoxSetRemoteDataSource dataSource ;
+  PillBoxSetRemoteDataSource dataSource;
 
   setUp(() {
-    dataSource = PillBoxSetRemoteDataSourceImpl();
+    // dataSource = PillBoxSetRemoteDataSourceImpl();
   });
 
   group('gets a PillBoxSet by dependent', () {
     test('returns a PillBoxSetModel', () async {
       // given
+      dataSource = PillBoxSetRemoteDataSourceImpl();
       final aDependent = 'Coda';
       final expectedPillBoxSet = fixture('pill_box_set.json');
       // when
@@ -25,6 +25,5 @@ void main() {
       // then
       expect(result, equals(expectedPillBoxSet));
     });
-
   });
 }
