@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:pill_pusher/features/schedule/domain/entities/pill.dart';
-import 'package:test/test.dart';
 
 void main() {
   final pill = Pill(name: "Zinc");
@@ -16,13 +16,11 @@ void main() {
   });
 
   group("Equatable", () {
-    test('props contains list of all properties that determine equality when constructed', ()
-    {
+    test('props contains list of all properties that determine equality when constructed', () {
       expect(pill.props, equals([pill.name]));
     });
 
-    test('stringify is turned on when constructed', ()
-    {
+    test('stringify is turned on when constructed', () {
       expect(pill.stringify, isTrue);
     });
   });
